Feat:  자동완성 검색바 공용 컴포넌트로 분리, 탐험페이지 코드 리팩토링
